Mise à jour
- Modification de l'utilisation des CT_Point. Les CT_Point sont à présent des Eigen::Vector3d (double !) et la classe CT_PointData est utilisé en interne pour stocker les points sous forme de Eigen::Vector3f (float)

- Suppression de l'accès au nuage de points global dans le PS_REPOSITORY. Utilisé maintenant la classe CT_PointAccessor dans le dossier ct_accessor

- Suppression de l'accès au point à partir d'un nuage d'index (constTAt et tAt). Utilisé plutôt un CT_PointIterator (pour l'accès en lecture) ou un CT_MutablePointIterator (pour l'accès en modification) dans le dossier ct_iterator

- Ajout de typedef pour les éléments retournés par le PS_REPOSITORY afin de raccourcir les noms des classes.

- Rédaction en cours d'une documentation sur le wiki pour expliquer l'utilisation/création des points.

ATTENTION ne pas faire cette mise à jour si vous ne voulez pas casser votre plugin qui devra subit quelques changement ! Attendre la fin de la rédaction de la documentation pour savoir comment modifier votre plugin.

git-svn-id: http://rdinnovation.onf.fr/svn/computree@291 0d0f61bb-bbfd-4940-b5ed-fa75979e1372

+++ pluginbase/actions/pb_actiondefineheightlayer.h
@@ -18,8 +18,8 @@ public:
 
     CT_Grid2DXY<int>    *_densityGrid;
     CT_Grid2DXY<float>  *_mnsGrid;
-    float               _zmin;
-    float               _zmax;
+    double              _zmin;
+    double              _zmax;
 };
 
 class PB_ActionDefineHeightLayer : public CT_AbstractActionForGraphicsView
@@ -28,7 +28,7 @@ class PB_ActionDefineHeightLayer : public CT_AbstractActionForGraphicsView
 public:
 
     PB_ActionDefineHeightLayer();
-    PB_ActionDefineHeightLayer(const CT_AbstractResult *result, const QString &densityGridModel, const QString &mnsGridModel, PB_ActionDefineHeightLayer_gridContainer *gridContainer, const QList<CT_Scene*> &list, float xmin, float ymin, float zmin, float xmax, float ymax, float zmax);
+    PB_ActionDefineHeightLayer(const CT_AbstractResult *result, const QString &densityGridModel, const QString &mnsGridModel, PB_ActionDefineHeightLayer_gridContainer *gridContainer, const QList<CT_Scene*> &list, double xmin, double ymin, double zmin, double xmax, double ymax, double zmax);
 
     void createGrids(float res);
 
@@ -60,7 +60,7 @@ public slots:
     void updateGraphics();
     void setSideView();
     void computeCrownProjection();
-    void updateZValues(float zmin, float zmax);
+    void updateZValues(double zmin, double zmax);
     void updateResolution(double resolution);
     void updateThreshold(int threshold);
